pyproject.toml
[tool.poetry]
name = "ubahnsiminstaller"
version = "1.1.0"
description = ""
authors = ["David Sn <divad.nnamtdeis@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
appdirs = "^1.4.4"
chardet = "^5.2.0"
psutil = "^6.0"
pillow = "^10.3.0"
pydantic = "^2.7.1"
pyinstaller = "^6.7.0"
pythonnet = "^3.0.3"
pywebview = "^5.1"
requests = "^2.32.2"
wmi = "^1.5.1"
sqlalchemy = "^2.0.31"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
